modularizing last commit
