+++ src/Dashboard/Dashboard.tsx
@@ -1,8 +1,8 @@
 import {useEffect, useState} from "react";
 import axiosInstance from "../Config/axiosInstance";
 import Logout from "../Components/Buttons/Logout";
-import UserOrgs from "../Components/UserOrgs";
 import CreateOrg from "../Components/CreateOrg";
+import ManageOrgs from "../Components/ManageOrgs";
 
 export default function Dashboard() {
     const [userInfo, setUserInfo] = useState<{ user_id: string; name: string; email: string } | null>(null);
@@ -31,7 +31,7 @@ export default function Dashboard() {
                         <p>Email: {userInfo.email}</p>
                     </div>
                 )}
-                <UserOrgs/>
+                <ManageOrgs/>
                 < Logout />
                 <CreateOrg/>
             </div>
